[DF-421] update win32 project.
Former-commit-id: 6a83e2fda79e7c942c3d90b2ac7fc5fb42131165

+++ Sources/Internal/Particles/ParticlePropertyLine.cpp
@@ -253,8 +253,7 @@ YamlNode* PropertyLineYamlWriter::WriteColorPropertyLineToYamlNode(YamlNode* par
 
 Vector4 PropertyLineYamlWriter::ColorToVector(const Color& color)
 {
-    return Vector4(int(color.r * 0xFF), int(color.g * 0xFF),
-                   int(color.b * 0xFF), int(color.a * 0xFF));
+    return Vector4(color.r * 0xFF, color.g * 0xFF, color.b * 0xFF, color.a * 0xFF);
 }
 
 }
